+++ AzLocal.UpdateManagement/Public/Get-AzLocalClusterInventory.ps1
@@ -298,7 +298,8 @@ function Get-AzLocalClusterInventory {
             # and throw a validation error for any cluster missing the tag.
             $ringTagValue = Get-TagValue -Tags $cluster.tags -Name 'UpdateRing'
             $windowTagValue = Get-TagValue -Tags $cluster.tags -Name $script:UpdateWindowTagName
-            $exclusionsTagValue = Get-TagValue -Tags $cluster.tags -Name $script:UpdateExclusionsTagName
+            $exclusionsTagValue = Get-TagValue -Tags $cluster.tags -Name $script:UpdateWindowExclusionsTagName
+            $excludedTagValue = Get-TagValue -Tags $cluster.tags -Name $script:UpdateExcludedTagName
             $sideloadedTagValue = Get-TagValue -Tags $cluster.tags -Name $script:UpdateSideloadedTagName
             $versionInProgressTagValue = Get-TagValue -Tags $cluster.tags -Name $script:UpdateVersionInProgressTagName
 
@@ -310,7 +311,8 @@ function Get-AzLocalClusterInventory {
                 UpdateRing              = if ($ringTagValue) { $ringTagValue } else { "" }
                 HasUpdateRingTag        = if ($ringTagValue) { "Yes" } else { "No" }
                 UpdateWindow            = if ($windowTagValue) { $windowTagValue } else { "" }
-                UpdateExclusions        = if ($exclusionsTagValue) { $exclusionsTagValue } else { "" }
+                UpdateWindowExclusions  = if ($exclusionsTagValue) { $exclusionsTagValue } else { "" }
+                UpdateExcluded          = if ($excludedTagValue) { $excludedTagValue } else { "" }
                 UpdateSideloaded        = if ($sideloadedTagValue) { $sideloadedTagValue } else { "" }
                 UpdateVersionInProgress = if ($versionInProgressTagValue) { $versionInProgressTagValue } else { "" }
                 ResourceId              = $cluster.id
@@ -335,7 +337,7 @@ function Get-AzLocalClusterInventory {
 
                 # Determine export format from file extension
                 $extension = [System.IO.Path]::GetExtension($ExportPath).ToLower()
-                $exportData = $inventory | Select-Object ClusterName, ResourceGroup, SubscriptionId, SubscriptionName, UpdateRing, HasUpdateRingTag, UpdateWindow, UpdateExclusions, UpdateSideloaded, UpdateVersionInProgress, ResourceId
+                $exportData = $inventory | Select-Object ClusterName, ResourceGroup, SubscriptionId, SubscriptionName, UpdateRing, HasUpdateRingTag, UpdateWindow, UpdateWindowExclusions, UpdateExcluded, UpdateSideloaded, UpdateVersionInProgress, ResourceId
                 
                 switch ($extension) {
                     '.json' {